updated plotter to plot metrics
